Finished Project 1 (with some debugging)
This project has two swordsmen fight Mortal Kombat-style, each with 10
HP.  There are three types of swords, Rock, Paper, and Scissors,
corresponding to the attacks in the classic childrens' game. A
successfully hit knocks off 1 HP.  There is a 30% chance of a critical
hit that knocks off 2 HP.

Once a swordsman's HP hits 0, there is a small animation and the other
swordsman wins.